refactor: 💡 change to pages and add alias
