+++ vite.config.js
@@ -5,4 +5,10 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   base: process.env.NODE_ENV === 'production' ? '/react-sample-vite-js/' : '/',
   plugins: [react()],
+  // alias
+  resolve: {
+    alias: {
+      '@': '/src',
+    },
+  },
 })
